pipeline/requirements.txt
pandas
yfinance
supabase
scikit-learn
numpy
ta
python-dotenv
tensorflow
joblib
# tensorflowjs is problematic on Windows due to uvloop dependency.
# We will install it separately with --no-deps or use a different conversion method.
